Replace aboutConfig.ts with a vue component for more granular page control
